+++ test/fixtures/lavas.config.js
@@ -7,10 +7,19 @@
 
 const path = require('path');
 const BUILD_PATH = path.resolve(__dirname, 'dist');
+// const t = JSON.stringify('test-var');
 
 module.exports = {
+    middleware: {
+        all: ['middleware1', 'middleware2']
+    },
     build: {
-        path: BUILD_PATH
+        path: BUILD_PATH,
+        defines: {
+            base: {
+                'DEFINE_TEST_VAR': '"test-var"'
+            }
+        }
     },
     entry: [
         {