+++ test/fixtures/middlewares/middleware1.js
@@ -1,0 +1,3 @@
+export default function ({route, store}) {
+    console.log('middleware1 is running...');
+}