+++ test/unit/route-manager.test.js
@@ -7,56 +7,68 @@
 
 import {join} from 'path';
 import test from 'ava';
-import LavasCore from '../../lib';
+import LavasCore from '../../dist/core';
 import {readFile} from 'fs-extra';
+import {syncConfig} from '../utils';
 
 let core;
 
 test.beforeEach('init', async t => {
     core = new LavasCore(join(__dirname, '../fixtures'));
-    await core.init('production');
+    await core.init('development', true);
 });
 
 /**
- * run the tests serially because they both modify .lavas/routes.js
+ * run the tests serially because they both modify .lavas/main/router.js
  *
  */
-test.serial('it should generate routes.js in .lavas directory', async t => {
-    await core.routeManager.buildRoutes();
+test.serial('it should generate main/router.js in .lavas directory', async t => {
+    await core.builder.routeManager.buildRoutes();
 
-    let content = await readFile(join(__dirname, '../fixtures/.lavas/routes.js'), 'utf8');
+    let content = await readFile(join(__dirname, '../fixtures/.lavas/main/router.js'), 'utf8');
 
-    t.true(content.indexOf('path: \'/detail/:id\'') > -1
+    t.true(content.indexOf('path: \':id\'') > -1
         && content.indexOf('name: \'detailId\'') > -1
         && content.indexOf('path: \'/\'') > -1
         && content.indexOf('name: \'index\'') > -1);
 });
 
+/**
+ * Use lavas.config.router to override some rules
+ * NOTE: routes.pattern apply to route.fullPath
+ *       rewrite.from apply to route.path
+ */
 test.serial('it should modify route objects based on router config', async t => {
 
-    Object.assign(core.config.router, {
+    syncConfig(core, Object.assign(core.config.router, {
+        rewrite: [
+            {
+                from: '/detail',
+                to: '/rewrite/detail'
+            }
+        ],
         routes: [
             {
                 pattern: /\/detail/,
                 lazyLoading: true,
                 chunkname: 'my-chunk',
-                path: '/detail/rewrite/:id',
                 meta: {
                     keepAlive: true
                 }
             }
         ]
-    });
+    }));
 
-    await core.routeManager.buildRoutes();
+    await core.builder.routeManager.buildRoutes();
 
-    let content = await readFile(join(__dirname, '../fixtures/.lavas/routes.js'), 'utf8');
+    let content = await readFile(join(__dirname, '../fixtures/.lavas/main/router.js'), 'utf8');
 
-    // Webpack code-splitting
-    t.true(content.indexOf('() => import(/* webpackChunkName: \"my-chunk\" */ \'@/pages/detail/_id.vue\');') > -1);
+    // Webpack code-splitting, merge Detail.vue & _id.vue into my-chunk
+    t.true(content.indexOf('() => import(/* webpackChunkName: \"my-chunk\" */\'@/pages/Detail.vue\');') > -1);
+    t.true(content.indexOf('() => import(/* webpackChunkName: \"my-chunk\" */\'@/pages/detail/_id.vue\');') > -1);
 
     // rewrite route path
-    t.true(content.indexOf('path: \'/detail/rewrite/:id\'') > -1);
+    t.true(content.indexOf('path: \'/rewrite/detail\'') > -1);
 
     // support route meta
     t.true(content.indexOf('meta: {"keepAlive":true}') > -1);
@@ -72,14 +84,15 @@ test.serial('it should modify route objects based on router config', async t => 
 // }
 
 // test.serial('it should generate routes.json in dist directory in prod mode', async t => {
-//     await core.routeManager.buildRoutes();
-
-//     let routes = core.routeManager.routes;
-
+//     let routeManager = core.builder.routeManager;
+//     await routeManager.buildRoutes();
+//
+//     let routes = routeManager.routes;
+//
 //     // regexp can't be serialized
 //     emptyRegExp(routes);
-
+//
 //     let savedRoutes = JSON.parse(await readFile(join(__dirname, '../fixtures/dist/routes.json'), 'utf8'));
-
-//     t.deepEqual(routes, savedRoutes)
+//
+//     t.deepEqual(routes, savedRoutes);
 // });